requirements.txt
keyboard
pydirectinput-rgx
python-dotenv
openai
pyttsx3
sounddevice
soundfile
numpy
elevenlabs
pytesseract
easyocr
mss
opencv-python
pillow
watchdog
dearpygui
colorlog
